requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
blinker==1.8.2
    # via flask
click==8.1.7
    # via flask
exceptiongroup==1.2.2
    # via pytest
flask==3.0.3
    # via -r requirements.in
iniconfig==2.0.0
    # via pytest
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via flask
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
packaging==24.1
    # via pytest
pluggy==1.5.0
    # via pytest
pytest==8.3.2
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
tomli==2.0.1
    # via pytest
werkzeug==3.0.4
    # via flask
